working on tree lists
