--- Tiralabra/nbproject/SmartYahtzee/src/smartyahtzee/AI/YahtzeeList.java
@@ -1,15 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package smartyahtzee.AI;
-
-/**
- *
- * @author Essi
- */
-public class YahtzeeList {
-    
-}
